refactor: move connection handling logic to resolver package
