--- pkg/gateway/internal/database/errors.go
@@ -1,16 +1,0 @@
-package database
-
-import (
-	"net/http"
-
-	"github.com/labstack/echo/v4"
-)
-
-var (
-	ErrInvalidCredentials = echo.NewHTTPError(http.StatusUnauthorized, "Invalid credentials")
-	ErrInvalidSubject     = echo.NewHTTPError(http.StatusBadRequest, "Invalid subject")
-	ErrInvalidUser        = echo.NewHTTPError(http.StatusBadRequest, "Invalid user")
-
-	ErrUserAlreadyExists = echo.NewHTTPError(http.StatusConflict, "User already exists")
-	ErrUserNotFound      = echo.NewHTTPError(http.StatusNotFound, "User not found")
-)
