+++ pkg/gateway/internal/database/models.go
@@ -1,5 +1,20 @@
 package database
 
+import (
+	"net/http"
+
+	"github.com/labstack/echo/v4"
+)
+
+var (
+	ErrInvalidCredentials = echo.NewHTTPError(http.StatusUnauthorized, "Invalid credentials")
+	ErrInvalidSubject     = echo.NewHTTPError(http.StatusBadRequest, "Invalid subject")
+	ErrInvalidUser        = echo.NewHTTPError(http.StatusBadRequest, "Invalid user")
+
+	ErrUserAlreadyExists = echo.NewHTTPError(http.StatusConflict, "User already exists")
+	ErrUserNotFound      = echo.NewHTTPError(http.StatusNotFound, "User not found")
+)
+
 type Session struct {
 	ID               string `json:"id"`
 	BrowserName      string `json:"browserName"`
@@ -8,6 +23,8 @@ type Session struct {
 	Platform         string `json:"platform"`
 	PlatformVersion  string `json:"platformVersion"`
 	DeviceModel      string `json:"deviceModel"`
+	UserHandle       string `json:"userHandle"`
+	VaultAddress     string `json:"vaultAddress"`
 }
 
 type User struct {
