+++ pkg/gateway/internal/database/sqlite.go
